Revert "Parameter guidelines init"
