+++ .gitignore
@@ -6,8 +6,5 @@ __pycache__/
 tex-tables/
 architecture/
 evaluation/
-models/
 main.py
 scheme/_neighbour_based.py
-parameter_guidelines/fingerprinted_data/
-archive/
