--- attacks/_additive_attack.py
@@ -1,1 +1,0 @@
-# todo
