+++ attacks/_subset_attack.py
@@ -1,7 +1,6 @@
 from attacks._base import Attack
 import time
 import random
-from datasets import Dataset
 
 
 class HorizontalSubsetAttack(Attack):
@@ -13,14 +12,12 @@ class HorizontalSubsetAttack(Attack):
     Runs the attack; gets a random subset of a dataset of size fraction*data_size
     fraction [0,1]
     """
-    def run(self, dataset, fraction, random_state=None):
-        if isinstance(dataset, Dataset):
-            dataset = dataset.get_dataframe()
+    def run(self, dataset, fraction):
         if fraction < 0 or fraction > 1:
             return None
 
         start = time.time()
-        subset = dataset.sample(frac=fraction, random_state=random_state)
+        subset = dataset.sample(frac=fraction)
         print("Subset attack runtime on " + str(int(fraction*len(dataset))) + " out of " + str(len(dataset)) +
               " entries: " + str(time.time()-start) + " sec.")
         return subset
