+++ datasets/_dataset.py
@@ -1,16 +1,14 @@
 from abc import ABC, abstractmethod
 import pandas as pd
 import os
-from sklearn.preprocessing import LabelEncoder, OneHotEncoder, StandardScaler
-import numpy as np
-from utilities import number_encode_features
+from sklearn.preprocessing import LabelEncoder
 
 
 class Dataset(ABC):
     """
     Abstract scheme for structuring the dataset within the toolbox
     """
-    def __init__(self, target_attribute, path=None, dataframe=None, primary_key_attribute=None, na_values=None):
+    def __init__(self, target_attribute, path=None, dataframe=None, primary_key_attribute=None):
         if path is None and dataframe is None:
             raise ValueError('Error defining a data set! Please provide either a path or a Pandas DataFrame to '
                              'instantiate a data set')
@@ -21,7 +19,7 @@ class Dataset(ABC):
         if self.path is not None and not isinstance(self.path, str):
             raise TypeError('Data set path must be a string value.')
         elif self.path is not None:
-            self.dataframe = pd.read_csv(self.path, na_values=na_values)
+            self.dataframe = pd.read_csv(self.path)
 
         if not isinstance(self.dataframe, pd.DataFrame):
             raise TypeError('Data frame must be type pandas.DataFrame')
@@ -160,64 +158,17 @@ class GermanCredit(Dataset):
         path = 'datasets/german_credit_full.csv'
         super().__init__(path=path, target_attribute='target', primary_key_attribute='Id')
 
-    def to_string(self):
-        return 'german_credit'
-
-    def clone(self):
-        clone = GermanCredit()
-        clone.set_target_attribute(self.target_attribute)
-        clone.set_dataframe(self.get_dataframe())
-        clone._set_primary_key(self.get_primary_key_attribute())
-        return clone
-
 
 class BreastCancerWisconsin(Dataset):
     def __init__(self):
         path = os.path.dirname(os.path.realpath(__file__)) + '/breast_cancer_wisconsin.csv'
         super().__init__(path=path, target_attribute='class', primary_key_attribute='sample-code-number')
 
-    def to_string(self):
-        return 'breast_cancer_wisconsin'
-
-    def clone(self):
-        clone = BreastCancerWisconsin()
-        clone.set_target_attribute(self.target_attribute)
-        clone.set_dataframe(self.get_dataframe())
-        clone._set_primary_key(self.get_primary_key_attribute())
-        return clone
-
 
 class Adult(Dataset):
     def __init__(self):
         path = 'datasets/adult.csv'
-        super().__init__(path=path, target_attribute='income', na_values='?')
-
-    def preprocessed(self, fp_data=None):
-        if fp_data is None:
-            preprocessed = self.clone()
-        # drop rows with missing values
-            preprocessed = preprocessed.dataframe.dropna().reset_index().drop('index', axis=1)
-        else:
-            preprocessed = fp_data
-
-        preprocessed, encoders = number_encode_features(preprocessed)
-        X = preprocessed.drop(self.target_attribute, axis=1)
-        y = preprocessed[self.target_attribute]
-
-        scaler = StandardScaler()
-        X = pd.DataFrame(scaler.fit_transform(X), columns=X.columns)
-        preprocessed = pd.concat([X, y], axis=1)
-        return preprocessed
-
-    def to_string(self):
-        return 'adult'
-
-    def clone(self):
-        clone = Adult()
-        clone.set_target_attribute(self.target_attribute)
-        clone.set_dataframe(self.get_dataframe())
-        clone._set_primary_key(self.get_primary_key_attribute())
-        return clone
+        super().__init__(path=path, target_attribute='income')
 
 
 class BreastCancer(Dataset):
@@ -225,67 +176,14 @@ class BreastCancer(Dataset):
         path = 'datasets/breast_cancer_full.csv'
         super().__init__(path=path, primary_key_attribute='Id', target_attribute='recurrence')
 
-    def to_string(self):
-        return 'breast_cancer'
-
-    def clone(self):
-        clone = BreastCancer()
-        clone.set_target_attribute(self.target_attribute)
-        clone.set_dataframe(self.get_dataframe())
-        clone._set_primary_key(self.get_primary_key_attribute())
-        return clone
-
 
 class Nursery(Dataset):
     def __init__(self):
         path = 'datasets/nursery_full.csv'
         super().__init__(path=path, primary_key_attribute='Id', target_attribute='target')
 
-    def to_string(self):
-        return 'nursery'
-
-    def clone(self):
-        clone = Nursery()
-        clone.set_target_attribute(self.target_attribute)
-        clone.set_dataframe(self.get_dataframe())
-        clone._set_primary_key(self.get_primary_key_attribute())
-        return clone
-
 
 class Mushrooms(Dataset):
     def __init__(self):
         path = 'datasets/mushroom_full.csv'
         super().__init__(path=path, primary_key_attribute='Id', target_attribute='target')
-
-    def to_string(self):
-        return 'mushrooms'
-
-    def clone(self):
-        clone = Mushrooms()
-        clone.set_target_attribute(self.target_attribute)
-        clone.set_dataframe(self.get_dataframe())
-        clone._set_primary_key(self.get_primary_key_attribute())
-        return clone
-
-
-def read_data(dataset, primary_key_attribute=None, target_attribute=None):
-    '''
-    Creates the instance of Dataset for given data.
-    :param dataset: string, pandas dataframe or Dataset
-    :param primary_key_attribute: name of the primary key attribute
-    :param target_attribute: name of the target attribute
-    :return: Dataset instance
-    '''
-    relation = None
-    if isinstance(dataset, str):  # assumed the path is given
-        relation = Dataset(path=dataset, target_attribute=target_attribute,
-                           primary_key_attribute=primary_key_attribute)
-    elif isinstance(dataset, pd.DataFrame):  # assumed the pd.DataFrame is given
-        relation = Dataset(dataframe=dataset, target_attribute=target_attribute,
-                           primary_key_attribute=primary_key_attribute)
-    elif isinstance(dataset, Dataset):
-        relation = dataset
-    else:
-        print('Wrong type of input data.')
-        exit()
-    return relation
