+++ requirements.txt
@@ -1,7 +1,7 @@
 bitstring==3.1.7
 matplotlib==3.3.4
 pyblake2==1.1.2
-astropy==4.2.1
+astropy==4.2
 joblib==1.0.1
 numpy==1.20.1
 pandas==1.2.3
