--- scheme/test/ak_scheme_2_32_0.csv
@@ -1,31 +1,0 @@
-Id,checking_account,duration,credit_hist,purpose,credit_amount,savings,employment_since,installment_rate,sex_status,debtors,residence_since,target
-0,A13,6,A34,A43,1169,A65,A75,4,A93,A101,4,1
-1,A12,48,A32,A43,5951,A61,A73,2,A92,A101,2,2
-2,A14,12,A34,A46,2096,A61,A74,2,A93,A101,3,0
-3,A11,42,A32,A42,7882,A61,A74,2,A93,A103,4,1
-4,A11,24,A30,A40,4870,A61,A73,3,A93,A101,4,2
-5,A14,36,A32,A46,9055,A65,A73,2,A93,A103,4,1
-6,A14,24,A32,A42,2835,A63,A75,3,A93,A101,4,1
-7,A12,36,A32,A42,6948,A61,A73,2,A93,A101,2,1
-8,A14,12,A32,A43,3059,A64,A74,2,A91,A101,4,1
-9,A11,30,A34,A40,5234,A61,A71,4,A94,A101,2,2
-10,A12,12,A32,A40,1295,A61,A72,3,A92,A101,1,2
-11,A11,48,A32,A49,4308,A61,A72,3,A92,A101,4,2
-12,A12,12,A32,A43,1567,A61,A73,1,A92,A101,1,1
-13,A11,24,A34,A40,1199,A61,A75,4,A93,A101,4,2
-14,A11,14,A32,A40,1403,A61,A73,2,A92,A101,4,1
-15,A11,24,A32,A43,1282,A62,A73,4,A92,A101,2,2
-16,A14,24,A34,A43,2424,A65,A75,4,A91,A101,4,1
-17,A11,30,A30,A49,8072,A65,A72,2,A93,A101,3,1
-18,A12,24,A32,A41,12579,A61,A75,4,A92,A101,2,2
-19,A14,24,A32,A43,3430,A63,A75,3,A93,A101,2,1
-20,A14,9,A34,A40,2134,A61,A73,4,A93,A101,4,1
-21,A11,6,A32,A43,2647,A63,A73,2,A93,A101,3,1
-22,A11,10,A34,A40,2241,A61,A73,1,A93,A101,3,1
-23,A12,12,A34,A41,1804,A62,A72,3,A93,A101,4,1
-24,A14,10,A34,A42,2069,A65,A73,2,A94,A101,1,1
-25,A11,6,A32,A42,1374,A61,A73,1,A93,A101,2,1
-26,A14,6,A30,A43,426,A61,A75,4,A94,A101,4,1
-27,A13,12,A31,A43,409,A64,A73,3,A92,A101,3,1
-28,A12,7,A32,A43,2415,A61,A73,3,A93,A103,2,1
-29,A11,60,A33,A49,6836,A61,A75,3,A93,A101,4,2
